add: codestyle, unified big dec usage

+++ src/main/java/ch/lueem/tradingbot/adapters/config/backtest/PortfolioConfig.java
@@ -1,7 +1,9 @@
 package ch.lueem.tradingbot.adapters.config.backtest;
 
+import java.math.BigDecimal;
+
 /**
  * Holds the configured portfolio settings for the backtest.
  */
-public record PortfolioConfig(double initialCash) {
+public record PortfolioConfig(BigDecimal initialCash) {
 }
